Introducing Core
This is my personal library with many methods that are extremely useful. Sort() takes a hashmap of any key type mapped to a double and returns a sorted arraylist where the keys are sorted based on the original double they were mapped to. There are many I/O methods that make reading from the terminal or a file instant by giving you an arraylist<String> containing each line from the stream. There are also many debugging helper functions including freeze() which stops the program and then continues when Enter is pressed.